refactor(workflow): extract plan database tests
